# main/docs/swagger.yaml
basePath: /api/
info:
  contact: {}
  description: 运维平台
  title: 运维平台
  version: 1.0.0
paths:
...
  /addMysqlBackupDetail/:
    post:
      description: 添加Mysql备份信息
      operationId: add-mysql-backup-detail
      parameters:
      - description: 数据库实例名字
        in: formData
        name: db_name
        required: true
        type: string
      - description: 实例端口号
        in: formData
        name: db_port
        required: true
        type: string
      - description: 备份服务器地址
        in: formData
        name: remote_server
        required: true
        type: string
      - description: 备份目录
        in: formData
        name: remote_dir
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已创建"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - mysqlBackup
...
  /allMysqlBackupInfo/:
    get:
      description: 获取全部Mysql备份信息
      operationId: get-all-mysql-backup-detail
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":[...]}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - mysqlBackup
  /allOtherBackupInfo/:
    get:
      description: 获取全部其他项目备份信息
      operationId: get-all-other-backup-detail
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":[...]}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - otherBackup
...
  /deleteAllMysqlBackupDetail/:
    delete:
      description: 删除全部Mysql备份信息
      operationId: delete-all-mysql-backup-detail
      parameters:
      - in: formData
        items:
          type: integer
        name: detail_ids
        type: array
      responses:
        "200":
          description: '{"code":200,"msg":"已全部删除"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - mysqlBackup
...
  /mysqlBackupDetail/{id}/:
    delete:
      description: 删除单条Mysql备份信息
      operationId: delete-mysql-backup-detail
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - mysqlBackup
...
  /serviceInit/mysql/:
    post:
      description: mysql初始化
      operationId: mysql-init
      parameters:
      - description: 初始化服务器
        in: formData
        name: ip
        required: true
        type: string
      - description: 数据库版本
        in: formData
        name: version
        required: true
        type: string
      - description: 实例端口号
        in: formData
        name: port
        required: true
        type: string
      - description: 实例名
        in: formData
        name: name
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已开始初始化"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - service
  /serviceInit/redis/:
    post:
      description: redis初始化
      operationId: redis-init
      parameters:
      - description: 初始化服务器
        in: formData
        name: ip
        required: true
        type: string
      - description: redis最大内存
        in: formData
        name: memsize
        required: true
        type: string
      - description: 实例端口号
        in: formData
        name: port
        required: true
        type: string
      - description: 实例名
        in: formData
        name: name
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已开始初始化"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - service
...
  /updateMysqlBackupDetail/:
    put:
      description: 更新mysql备份信息
      operationId: update-mysql-backup-detail
      parameters:
      - description: 数据库实例名
        in: formData
        name: db_name
        required: true
        type: string
      - description: 端口号
        in: formData
        name: db_port
        required: true
        type: string
      - description: 备份服务器地址
        in: formData
        name: remote_server
        required: true
        type: string
      - description: 备份目录
        in: formData
        name: remote_dir
        required: true
        type: string
      - description: 备份信息
        in: formData
        name: backup_log
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已更新"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - mysqlBackup
  /updateOtherBackupDetail/:
    put:
      description: 更新其他项目备份信息
      operationId: update-other-backup-detail
      parameters:
      - description: 项目名字
        in: formData
        name: project_name
        required: true
        type: string
      - description: 备份服务器地址
        in: formData
        name: remote_server
        required: true
        type: string
      - description: 备份目录
        in: formData
        name: remote_dir
        required: true
        type: string
      - description: 备份信息
        in: formData
        name: backup_log
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"已更新"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - otherBackup
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"